# open-api-test/spring-server/src/main/java/org/openapitools/api/UsersApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ResponseMessage;
import org.openapitools.model.UserRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-08-08T12:32:36.175772600+09:00[Asia/Tokyo]", comments = "Generator version: 7.14.0")

@Validated
@Tag(name = "Users", description = "the Users API")
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    public static final String PATH_USERS_POST = "/users";
    /**
     * POST /users : ユーザ登録
     * 新しいユーザを登録します
     *
     * @param userRequest  (required)
     * @return ユーザ登録成功 (status code 200)
     *         or リクエスト不正 (status code 400)
     */
    @Operation(
        operationId = "usersPost",
        summary = "ユーザ登録",
        description = "新しいユーザを登録します",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "ユーザ登録成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ResponseMessage.class))
            }),
            @ApiResponse(responseCode = "400", description = "リクエスト不正", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ResponseMessage.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = UsersApi.PATH_USERS_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ResponseMessage> usersPost(
        @Parameter(name = "UserRequest", description = "", required = true) @Valid @RequestBody UserRequest userRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"操作が成功しました\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"操作が成功しました\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_USERS_USER_ID_DELETE = "/users/{userId}";
    /**
     * DELETE /users/{userId} : ユーザ削除
     * 指定したユーザIDのユーザを削除します
     *
     * @param userId 削除対象のユーザID (required)
     * @return ユーザ削除成功 (status code 200)
     *         or ユーザが存在しない (status code 404)
     */
    @Operation(
        operationId = "usersUserIdDelete",
        summary = "ユーザ削除",
        description = "指定したユーザIDのユーザを削除します",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "ユーザ削除成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ResponseMessage.class))
            }),
            @ApiResponse(responseCode = "404", description = "ユーザが存在しない", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ResponseMessage.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = UsersApi.PATH_USERS_USER_ID_DELETE,
        produces = { "application/json" }
    )
    
    default ResponseEntity<ResponseMessage> usersUserIdDelete(
        @Parameter(name = "userId", description = "削除対象のユーザID", required = true, in = ParameterIn.PATH) @PathVariable("userId") String userId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"操作が成功しました\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"message\" : \"操作が成功しました\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}