# open-api-test/ts-client/models/ResponseMessage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * User Management API
 * ユーザの登録と削除を行うAPI
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ResponseMessage
 */
export interface ResponseMessage {
    /**
     * 結果メッセージ
     * @type {string}
     * @memberof ResponseMessage
     */
    message?: string;
}

/**
 * Check if a given object implements the ResponseMessage interface.
 */
export function instanceOfResponseMessage(value: object): value is ResponseMessage {
    return true;
}

export function ResponseMessageFromJSON(json: any): ResponseMessage {
    return ResponseMessageFromJSONTyped(json, false);
}

export function ResponseMessageFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseMessage {
    if (json == null) {
        return json;
    }
    return {
        
        'message': json['message'] == null ? undefined : json['message'],
    };
}

export function ResponseMessageToJSON(json: any): ResponseMessage {
    return ResponseMessageToJSONTyped(json, false);
}

export function ResponseMessageToJSONTyped(value?: ResponseMessage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
    };
}