# open-api-test/ts-client/models/UserRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * User Management API
 * ユーザの登録と削除を行うAPI
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UserRequest
 */
export interface UserRequest {
    /**
     * ユーザ名
     * @type {string}
     * @memberof UserRequest
     */
    username: string;
    /**
     * メールアドレス
     * @type {string}
     * @memberof UserRequest
     */
    email: string;
}

/**
 * Check if a given object implements the UserRequest interface.
 */
export function instanceOfUserRequest(value: object): value is UserRequest {
    if (!('username' in value) || value['username'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    return true;
}

export function UserRequestFromJSON(json: any): UserRequest {
    return UserRequestFromJSONTyped(json, false);
}

export function UserRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'],
        'email': json['email'],
    };
}

export function UserRequestToJSON(json: any): UserRequest {
    return UserRequestToJSONTyped(json, false);
}

export function UserRequestToJSONTyped(value?: UserRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'email': value['email'],
    };
}